PayPal payment integration

+++ store/models.py
@@ -1,5 +1,6 @@
 import email
 from email.policy import default
+from unicodedata import decimal
 from django.db import models
 from django.contrib.auth.models import User
 
@@ -13,7 +14,7 @@ class Customer(models.Model):
 
 class Product(models.Model):
     name = models.CharField(max_length=200, null=True)
-    price = models.FloatField()
+    price = models.DecimalField(max_digits=7, decimal_places=2)
     digital = models.BooleanField(default=False, null=True, blank=True)
     image = models.ImageField(null=True, blank=True)
